查询除入库全部完成成功

+++ src/main/java/cn/cvs/pojo/TAddress.java
@@ -14,7 +14,7 @@ public class TAddress implements Serializable {
     /**
      * 主键ID
      */
-    private Long id;
+    private Integer id;
     /**
      * 联系人姓名
      */
@@ -34,7 +34,7 @@ public class TAddress implements Serializable {
     /**
      * 创建者
      */
-    private Long createdUserId;
+    private Integer createdUserId;
     /**
      * 创建时间
      */
@@ -42,7 +42,7 @@ public class TAddress implements Serializable {
     /**
      * 修改者
      */
-    private Long updatedUserId;
+    private Integer updatedUserId;
     /**
      * 修改时间
      */
@@ -50,14 +50,14 @@ public class TAddress implements Serializable {
     /**
      * 用户ID
      */
-    private Long userid;
+    private Integer userid;
 
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -93,11 +93,11 @@ public class TAddress implements Serializable {
         this.tel = tel;
     }
 
-    public Long getCreatedUserId() {
+    public Integer getCreatedUserId() {
         return createdUserId;
     }
 
-    public void setCreatedUserId(Long createdUserId) {
+    public void setCreatedUserId(Integer createdUserId) {
         this.createdUserId = createdUserId;
     }
 
@@ -109,11 +109,11 @@ public class TAddress implements Serializable {
         this.createdTime = createdTime;
     }
 
-    public Long getUpdatedUserId() {
+    public Integer getUpdatedUserId() {
         return updatedUserId;
     }
 
-    public void setUpdatedUserId(Long updatedUserId) {
+    public void setUpdatedUserId(Integer updatedUserId) {
         this.updatedUserId = updatedUserId;
     }
 
@@ -125,11 +125,11 @@ public class TAddress implements Serializable {
         this.updatedTime = updatedTime;
     }
 
-    public Long getUserid() {
+    public Integer getUserid() {
         return userid;
     }
 
-    public void setUserid(Long userid) {
+    public void setUserid(Integer userid) {
         this.userid = userid;
     }
